requirements.txt
asgiref==3.3.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1.6
django-cors-headers==3.7.0
django-on-heroku==1.1.2
django-rest-swagger==2.2.0
djangorestframework==3.12.2
drf-spectacular==0.13.1
gunicorn==20.0.4
httpie==2.3.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==1.1.1
openapi-codec==1.3.2
psycopg2-binary==2.8.6
Pygments==2.7.4
PyMySQL==1.0.2
pyrsistent==0.17.3
PySocks==1.7.1
python-decouple==3.4
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
requests-toolbelt==0.9.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
whitenoise==5.2.0
